+++ nodocliente.h
@@ -1,5 +1,6 @@
 #ifndef _INCLUDE_NODO_CLIENTE_H_
 #define _INCLUDE_NODO_CLIENTE_H_
+#include <iostream>     // std::cout
 
 using namespace std;
 
@@ -10,18 +11,18 @@ class NodoCliente{
 		int idCliente;
 		double posicionx;
 		double posiciony;
-		int demanda;
+		double demanda;
 
 	public:
 		NodoCliente();
-		NodoCliente(int id, double x, double y, int d);
+		NodoCliente(int id, double x, double y, double d);
 		int getId();
 		void setId(int id);
 		double getCoordenadaX();
 		double getCoordenadaY();
 		void setPosicion(double x, double y);
-		int getDemanda();
-		void setDemanda(int d);
+		double getDemanda();
+		void setDemanda(double d);
 
 
 };
